feat: move typechain things in src/contracts
