+++ scripts/postinstall.sh
@@ -22,4 +22,4 @@ fi
 git config --local include.path '../.gitconfig'
 
 yarn patch-package
-yarn typechain src/state/slices/opportunitiesSlice/resolvers/uniV2/abis/*.json --target=ethers-v5 --out-dir ./src/state/slices/opportunitiesSlice/resolvers/uniV2/contracts
+yarn typechain src/contracts/abis/*.json --target=ethers-v5 --out-dir ./src/contracts/__generated
