+++ src/contracts/__generated/ERC20ABI.ts
