+++ src/contracts/__generated/IUniswapV2Pair.ts
