+++ src/contracts/__generated/IUniswapV2Router02.ts
