+++ src/contracts/__generated/common.ts
