+++ src/contracts/__generated/index.ts
