+++ src/plugins/foxPage/utils.ts
@@ -2,13 +2,13 @@ import type { Contract } from '@ethersproject/contracts'
 import type { AssetId } from '@shapeshiftoss/caip'
 import { fromAssetId } from '@shapeshiftoss/caip'
 import type { TokenAmount } from '@uniswap/sdk'
+import type { IUniswapV2Pair } from 'contracts/__generated'
+import { getOrCreateContract } from 'contracts/contractManager'
 import memoize from 'lodash/memoize'
 import type { BN } from 'lib/bignumber/bignumber'
 import { bn, bnOrZero } from 'lib/bignumber/bignumber'
 import { logger } from 'lib/logger'
 import type { foxEthLpContractAddress } from 'state/slices/opportunitiesSlice/constants'
-import { getOrCreateContract } from 'state/slices/opportunitiesSlice/resolvers/uniV2/contractManager'
-import type { IUniswapV2Pair } from 'state/slices/opportunitiesSlice/resolvers/uniV2/contracts'
 
 import { TRADING_FEE_RATE } from './const'
